Use driver for remote protocol introduced in da628be8.
The remote protocol was calling into the Storage object but this required some translation which will get more awkward as time goes by.

Instead, call directly into the local driver so the communication is directly driver to driver.  This still requires resolving the path and may eventually have more duplication with the Storage object methods but it seems the right thing to do.