Ignore write errors when the ls command is writing to stdout.
It's possible (even likely) that the ls output is being piped to something like head which will exit when it gets what it needs and leave us writing to a broken pipe.

It would be better to just ignore the broken pipe error but currently we don't store system error codes.